Refactor DownloadPage component to utilize new modular components for improved structure and readability. Introduce DownloadTabs, GetInvolved, and ExportCompliance components, enhancing maintainability and separation of concerns.
